# api/openapi.yaml
openapi: 3.0.3
info:
  title: DispScenario Analyst API
  version: 0.1.0
  description: Contract for the complete v2 replacement.
servers:
  - url: http://localhost:8787
security:
  - BearerAuth: []
paths:
...
  /v1/recordings/uploads:
    post:
      operationId: createRecordingUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUploadRequest"
      responses:
        "201":
          description: Upload session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadSession"
  /v1/recordings/{recordingId}/uploads/complete:
    post:
      operationId: completeRecordingUpload
      parameters:
        - $ref: "#/components/parameters/RecordingId"
      responses:
        "200":
          description: Completed recording
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"
  /v1/recordings/{recordingId}:
    delete:
      operationId: deleteRecording
      parameters:
        - $ref: "#/components/parameters/RecordingId"
      responses:
        "204":
          description: Recording deleted
  /v1/recordings/{recordingId}/playback:
    get:
      operationId: getRecordingPlayback
      parameters:
        - $ref: "#/components/parameters/RecordingId"
      responses:
        "200":
          description: Signed playback URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignedUrl"
  /v1/recordings/{recordingId}/analysis-runs:
    post:
      operationId: createAnalysisRun
      parameters:
        - $ref: "#/components/parameters/RecordingId"
      responses:
        "202":
          description: Analysis accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRun"
  /v1/analysis-runs:
    get:
      operationId: listAnalysisRuns
      parameters:
        - name: recordingId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Analysis runs
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AnalysisRun"
  /v1/analysis-runs/{runId}:
    get:
      operationId: getAnalysisRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Analysis run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRun"
  /v1/analysis-runs/{runId}/cancel:
    post:
      operationId: cancelAnalysisRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Cancelled run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRun"
  /v1/analysis-runs/{runId}/retry:
    post:
      operationId: retryAnalysisRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "202":
          description: Retried run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRun"
  /v1/recordings/{recordingId}/evidence/{timestampMs}:
    get:
      operationId: getEvidenceFrame
      parameters:
        - $ref: "#/components/parameters/RecordingId"
        - name: timestampMs
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
      responses:
        "307":
          description: Redirect to signed evidence image
  /v1/recordings/{recordingId}/events:
    get:
      operationId: listRecordingEvents
      parameters:
        - $ref: "#/components/parameters/RecordingId"
      responses:
        "200":
          $ref: "#/components/responses/JsonArray"
  /v1/recordings/{recordingId}/raw-events:
    get:
      operationId: listRecordingRawEvents
      parameters:
        - $ref: "#/components/parameters/RecordingId"
      responses:
        "200":
          $ref: "#/components/responses/JsonArray"
  /v1/recordings/{recordingId}/quality-issues:
    get:
      operationId: listRecordingQualityIssues
      parameters:
        - $ref: "#/components/parameters/RecordingId"
      responses:
        "200":
          $ref: "#/components/responses/JsonArray"
  /v1/recordings/{recordingId}/analysis:
    get:
      operationId: getRecordingAnalysis
      parameters:
        - $ref: "#/components/parameters/RecordingId"
      responses:
        "200":
          $ref: "#/components/responses/JsonObject"
  /v1/recordings/{recordingId}/scenario-map:
    get:
      operationId: getRecordingScenarioMap
      parameters:
        - $ref: "#/components/parameters/RecordingId"
      responses:
        "200":
          $ref: "#/components/responses/JsonObject"
  /v1/recordings/{recordingId}/rebuild:
    post:
      operationId: rebuildRecordingAnalysis
      parameters:
        - $ref: "#/components/parameters/RecordingId"
      responses:
        "200":
          $ref: "#/components/responses/JsonObject"
  /v1/recordings/{recordingId}/renormalize:
    post:
      operationId: renormalizeRecordingAnalysis
      parameters:
        - $ref: "#/components/parameters/RecordingId"
      responses:
        "200":
          $ref: "#/components/responses/JsonObject"
  /v1/recordings/{recordingId}/boundary-review:
    post:
      operationId: reviewRecordingBoundaries
      parameters:
        - $ref: "#/components/parameters/RecordingId"
      responses:
        "200":
          $ref: "#/components/responses/JsonObject"
  /v1/recordings/{recordingId}/events/{eventId}:
    patch:
      operationId: updateRecordingEvent
      parameters:
        - $ref: "#/components/parameters/RecordingId"
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
      responses:
        "200":
          $ref: "#/components/responses/JsonObject"
  /v1/recordings/{recordingId}/quality-issues/{issueId}:
    patch:
      operationId: updateRecordingQualityIssue
      parameters:
        - $ref: "#/components/parameters/RecordingId"
        - $ref: "#/components/parameters/IssueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
      responses:
        "200":
          $ref: "#/components/responses/JsonObject"
  /v1/recordings/{recordingId}/qa/complete:
    post:
      operationId: completeRecordingQA
      parameters:
        - $ref: "#/components/parameters/RecordingId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
      responses:
        "200":
          $ref: "#/components/responses/JsonObject"
  /v1/recordings/{recordingId}/exports/{kind}.{format}:
    get:
      operationId: exportRecording
      parameters:
        - $ref: "#/components/parameters/RecordingId"
        - name: kind
          in: path
          required: true
          schema:
            type: string
            enum: [timeline, report]
        - name: format
          in: path
          required: true
          schema:
            type: string
            enum: [json, csv]
      responses:
        "200":
          description: Export file
...
  /v1/reports/{reportId}:
    get:
      operationId: getReport
      parameters:
        - $ref: "#/components/parameters/ReportId"
      responses:
        "200":
          $ref: "#/components/responses/JsonObject"
  /v1/ground-truth/import:
    post:
      operationId: importGroundTruth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
      responses:
        "201":
          $ref: "#/components/responses/JsonObject"
...
  /v1/settings/gemini-credential:
    get:
      operationId: getGeminiCredential
      responses:
        "200":
          description: Credential status. The secret is never returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderCredentialStatus"
    put:
      operationId: updateGeminiCredential
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [apiKey]
              properties:
                apiKey:
                  type: string
                  minLength: 20
                  writeOnly: true
      responses:
        "200":
          description: Credential saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderCredentialStatus"
    delete:
      operationId: deleteGeminiCredential
      responses:
        "204":
          description: Credential deleted.
  /v1/settings/known-scenarios/{code}:
    patch:
      operationId: updateKnownScenario
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
      responses:
        "200":
          $ref: "#/components/responses/JsonObject"
  /v1/settings/boundary-rules/{ruleId}:
    patch:
      operationId: updateBoundaryRule
      parameters:
        - name: ruleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonObject"
      responses:
        "200":
          $ref: "#/components/responses/JsonObject"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RecordingId:
      name: recordingId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IssueId:
      name: issueId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ReportId:
      name: reportId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    JsonObject:
      description: JSON object
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JsonObject"
    JsonArray:
      description: JSON array
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/JsonObject"
  schemas:
    ProviderCredentialStatus:
      type: object
      required: [provider, configured]
      properties:
        provider:
          type: string
          example: gemini
        configured:
          type: boolean
        lastFour:
          type: string
          minLength: 4
          maxLength: 4
        updatedAt:
          type: string
          format: date-time
    JsonObject:
      type: object
      additionalProperties: true
...
    Recording:
      type: object
      required:
        - id
        - originalName
        - mimeType
        - sizeBytes
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        originalName:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
          format: int64
        durationSec:
          type: number
          format: double
          nullable: true
        status:
          type: string
          enum: [PENDING_UPLOAD, UPLOADED, PROCESSING, ANALYZED, FAILED]
        createdAt:
          type: string
          format: date-time
...
    CreateUploadRequest:
      type: object
      required: [originalName, mimeType, sizeBytes]
      properties:
        originalName:
          type: string
          minLength: 1
          maxLength: 512
        mimeType:
          type: string
          enum: [video/webm, video/mp4]
        sizeBytes:
          type: integer
          format: int64
          minimum: 1
          maximum: 2147483648
    UploadSession:
      type: object
      required: [recording, uploadUrl, expiresAt]
      properties:
        recording:
          $ref: "#/components/schemas/Recording"
        uploadUrl:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
    SignedUrl:
      type: object
      required: [url, expiresAt]
      properties:
        url:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
    AnalysisRun:
      type: object
      required:
        - id
        - recordingId
        - status
        - provider
        - promptVersion
        - normalizationVersion
        - groupingVersion
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        recordingId:
          type: string
          format: uuid
        status:
          type: string
          enum: [QUEUED, PROCESSING, NORMALIZING, COMPLETED, FAILED, CANCELLED]
        provider:
          type: string
        model:
          type: string
          nullable: true
        promptVersion:
          type: string
        normalizationVersion:
          type: string
        groupingVersion:
          type: string
        error:
          type: string
          nullable: true
        inputTokens:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        outputTokens:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        thinkingTokens:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        totalTokens:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        estimatedCostUsd:
          type: number
          format: double
          minimum: 0
          nullable: true
        pricingVersion:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time